requirements.txt
langchain
streamlit
pypdf2
python-docx
